rename Hardware namespace to HardwareControlelrs at RfidControllers app
https://github.com/Not-Endurance/not-timing-system/issues/281
